pyproject.toml
[project]
name = "backend-takehome"
version = "0.1.0"
description = ""
authors = [
    {name = "gulatikanan",email = "kanangulati7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "biopython (>=1.85,<2.0)",
    "rich (>=14.0.0,<15.0.0)",
    "typing-extensions (>=4.13.1,<5.0.0)"
]

[tool.poetry]
packages = [{include = "backend_takehome", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "backend_takehome.cli:main"


